Fixed minor bug where the filter full message was being reported as the dust collector was shutting down due to a signed/unsigned comparison.

+++ DCController/DustCollector.cpp
@@ -509,7 +509,7 @@ void DustCollector::CheckFilter(void)
 					*	If the average is greater than or equal to the dirty pressure THEN
 					*	set the status to filter full, start flasher, send message.
 					*/
-					if (deltaAverage >= mGateSets.CurrentDirtyPressure())
+					if (deltaAverage >= (int32_t)mGateSets.CurrentDirtyPressure())
 					{
 						mStatus = eFilterFull;
 						mFaultAcknowledged = false;
